requirements.txt
flask==2.0.1
flask-cors==3.0.10
rembg==2.0.50
Pillow==10.0.0
python-multipart==0.0.6